Small issues Correction
Correcting small issues.

On the Docs Blueprint Browser page, I said I selected the zealot when the avatar was selected.

On the blueprint Browser control, I wasn't always changing the selected type when the tree view was clicked. The full tree view was taking precedence over the search tree view.