making border systems clickable
In the System Info page, I am making the border systems clickable. This allows the user to switch to that system without having to go back to the grid and finding it.

I am also adding color coding to the system so that you can easily distinguish high sec, low sec, and null sec.