Issue #8 - Changing cache time of Market Orders
Changing the cache time of market orders to be 15 minutes instead of 5.

This will result in price updates taking a little longer, but it will improve the screen to screen performance of the app by not having to call ESI every time.

Also added the code that could get all orders in an entire region but commented it out because the length of time it was taking was significantly longer than getting items on a per type and region/location basis.