#11 Blueprint Store for Build Plans
Adding a blueprint store for the build plan control. This allows a user to set the ME/TE and Max Runs for reactions and blueprints involved in the creation of the item.

This allows for more accurate job costs, material costs, and build time.

still some work that needs to be done, but this is getting to be good.