Fixing issues with background workers
If a background worker was executing and I tried to wait on it to finish and then restart it, it was hanging the application.

I removed the tab stop property on all tree views. If the user searched, hit tab a couple times, it activated the tree view for the first item in the list. Then, if they clicked on an item in the tree view, it was causing a hangup with the above issue.