Performance improvements
Focused on performance improvements of both startup and build plans control.

redid the main screen to be flat buttons instead of the images. Scales better that way. Also less overhead.

fixed the background workers in build plan to support cancellation so that when you switch from one to the next, it cancels and reloads instead of hanging.

changed the build plan combo to reset the label values.