#50 - Fixing DPI Issue with build plans
DPI differences from one machine to the next was causing elements to shift and not be visible.

Changed the auto size mode to true on a few screens this happened on. Made sure elemtns were in line.

Tested by changing resolution on my dev machine to try to simulate different machines