Refactor Common Helper Code
Updating the common helper code to be agnostic as to how it gets the data for it's methods. It no longer relies on top level classes and instead only takes in the values that it cares about. This leaves the decision of where to get the data up to the parent ( reaction vs manufacturing, structure profile vs indy settings )